feat(saml): xcpd valid, mirth rejecting signature
Refs: #1667
