+++ packages/utils/src/saml/server.ts
@@ -1,5 +1,9 @@
 import express from "express";
 import { json, Request, Response } from "express";
+import axios from "axios";
+import fs from "fs";
+import https from "https";
+
 import { createSoapEnvelope } from "@metriport/core/external/carequality/saml/xcpd/envelope";
 import { verifyXmlSignatures } from "@metriport/core/external/carequality/saml/security/verify";
 import {
@@ -32,8 +36,13 @@ app.post("/xcpd", async (req: Request, res: Response) => {
       x509CertPem
     );
     console.log("Signatures verified: ", verified);
+    fs.writeFileSync("./temp.xml", signedTimestampAndEnvelope.getSignedXml());
+    const response = await sendSignedXml(
+      signedTimestampAndEnvelope.getSignedXml(),
+      req.body.gateway.url
+    );
 
-    res.type("application/xml").send(signedTimestampAndEnvelope.getSignedXml());
+    res.type("application/xml").send(response);
   } catch (error) {
     console.error(error);
     res.status(500).send({ detail: "Internal Server Error" });
@@ -43,3 +52,35 @@ app.post("/xcpd", async (req: Request, res: Response) => {
 app.listen(port, () => {
   console.log(`Server running at http://localhost:${port}`);
 });
+
+async function sendSignedXml(signedXml: string, url: string): Promise<string> {
+  const certFilePath = "./tempCert.pem";
+  const keyFilePath = "./tempKey.pem";
+  fs.writeFileSync(certFilePath, x509CertPem);
+  fs.writeFileSync(keyFilePath, privateKey);
+
+  try {
+    const agent = new https.Agent({
+      rejectUnauthorized: false,
+      cert: fs.readFileSync(certFilePath),
+      key: fs.readFileSync(keyFilePath),
+    });
+
+    const response = await axios.post(url, signedXml, {
+      headers: {
+        "Content-Type": "application/soap+xml;charset=UTF-8",
+        "Cache-Control": "no-cache",
+      },
+      httpsAgent: agent,
+    });
+
+    return response.data;
+  } catch (error) {
+    console.error("Failed to send signed XML:", error);
+    throw error;
+  } finally {
+    // Clean up the temporary files
+    fs.unlinkSync(certFilePath);
+    fs.unlinkSync(keyFilePath);
+  }
+}
